+++ retry/sql.go
@@ -8,6 +8,7 @@ import (
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/stack"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xcontext"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
+	budget "github.com/ydb-platform/ydb-go-sdk/v3/retry/budget"
 	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
 )
 
@@ -142,12 +143,14 @@ func DoTx(ctx context.Context, db *sql.DB, op func(context.Context, *sql.Tx) err
 		}
 		attempts = 0
 	)
-	if tracer, has := db.Driver().(interface {
+	if d, has := db.Driver().(interface {
 		TraceRetry() *trace.Retry
+		RetryBudget() budget.Budget
 	}); has {
-		options.retryOptions = append(options.retryOptions, nil)
-		copy(options.retryOptions[1:], options.retryOptions)
-		options.retryOptions[0] = WithTrace(tracer.TraceRetry())
+		options.retryOptions = append(options.retryOptions, nil, nil)
+		copy(options.retryOptions[2:], options.retryOptions)
+		options.retryOptions[0] = WithTrace(d.TraceRetry())
+		options.retryOptions[1] = WithBudget(d.RetryBudget())
 	}
 	for _, opt := range opts {
 		if opt != nil {