+++ tests/slo/database/sql/storage.go
@@ -9,6 +9,7 @@ import (
 
 	ydb "github.com/ydb-platform/ydb-go-sdk/v3"
 	"github.com/ydb-platform/ydb-go-sdk/v3/retry"
+	"github.com/ydb-platform/ydb-go-sdk/v3/retry/budget"
 	"github.com/ydb-platform/ydb-go-sdk/v3/table"
 	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
 
@@ -74,11 +75,18 @@ type Storage struct {
 	dropQuery   string
 	upsertQuery string
 	selectQuery string
+	retryBudget interface {
+		budget.Budget
+
+		Stop()
+	}
 }
 
 func NewStorage(ctx context.Context, cfg *config.Config, poolSize int) (s *Storage, err error) {
 	ctx, cancel := context.WithTimeout(ctx, time.Minute*5) //nolint:gomnd
 	defer cancel()
+
+	retryBudget := budget.Limited(int(float64(poolSize) * 0.1)) //nolint:gomnd
 
 	s = &Storage{
 		cfg: cfg,
@@ -87,11 +95,13 @@ func NewStorage(ctx context.Context, cfg *config.Config, poolSize int) (s *Stora
 		dropQuery:   fmt.Sprintf(dropTemplate, cfg.Table),
 		upsertQuery: fmt.Sprintf(upsertTemplate, cfg.Table),
 		selectQuery: fmt.Sprintf(selectTemplate, cfg.Table),
+		retryBudget: retryBudget,
 	}
 
 	s.cc, err = ydb.Open(
 		ctx,
 		s.cfg.Endpoint+s.cfg.DB,
+		ydb.WithRetryBudget(retryBudget),
 	)
 	if err != nil {
 		return nil, fmt.Errorf("ydb.Open error: %w", err)
@@ -224,6 +234,8 @@ func (s *Storage) dropTable(ctx context.Context) error {
 }
 
 func (s *Storage) close(ctx context.Context) error {
+	s.retryBudget.Stop()
+
 	if err := ctx.Err(); err != nil {
 		return err
 	}